pollers/pyproject.toml
[project]
name = "wapprepollers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Julius Laitala", email = "julius.laitala@brightlyworks.com" },
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp[speedups]>=3.11.18",
    "python-socketio[asyncio-client]~=4.6",
    "valkey[libvalkey]>=6.1.0",
    "wapprecommon",
    "websockets>=15.0.1",
]

[project.scripts]
wapprepollers = "wapprepollers:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
wapprecommon = { path = "../common", editable = true }

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "E",   # Pycodestyle errors
    "W",   # Pycodestyle warnings
    "N",   # pep8-naming
    "D",   # Pydocstyle
    "RUF", # Ruff-specifics
]
ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
    "D413", # Missing blank line after last section
]

[tool.mypy]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["socketio"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy>=1.15.0",
]
